requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
python-decouple>=3.8
Faker>=20.0.0
python-slugify>=8.0.0
drf-spectacular>=0.26.0
django-filter>=23.0
django-cors-headers>=4.3.0
django-extensions>=3.2.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0